pyproject.toml
[project]
name = "autoclaimopt"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
]

[tool.uv]
dev-dependencies = []
